requirements/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
clickhouse-driver==0.2.6
clickhouse-pool==0.5.3
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.8.0rc2
deprecation==2.1.0
Django==5.0.3
django-allauth==0.61.1
django-clickhouse-backend==1.1.6
django-cors-headers==4.3.1
django-email-verification==0.3.3
djangorestframework==3.15.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.0
websocket-client==1.7.0
zope.interface==6.2
